support evm version: Shanghai! (#85)
* chore: tmp

* feat: support push0

* feat: supoort shanghai

* support evm version: shanghai
* Supplant DIFFICULTY opcode with PREVRANDAO
* fix test

* chore: bump to 0.4.0

* chore: downgrade cita_trie

* chore: bump to 0.4.1

* fix: push0 lookup failed

* chore: bump to 0.4.2

+++ src/evm/common.rs
@@ -74,7 +74,7 @@ pub fn rpad(slice: Vec<u8>, n: usize) -> Vec<u8> {
         let mut padded: Vec<u8> = Vec::with_capacity(n);
         let mut part1 = slice;
         padded.append(&mut part1);
-        let mut part2 = vec![0; n as usize - slice_len];
+        let mut part2 = vec![0; n - slice_len];
         padded.append(&mut part2);
         padded
     }
